+++ test.csv
@@ -1,0 +1,9 @@
+obj_id,time,feat_name,feat_start,feat_end,vel,vel_err,vel_samperr,pew,pew_err,pew_samperr,calc_area,calc_area_err,calc_area_samperr,notes
+2004ef,2453262.61,pW2,3927.684,4092.8837,9691.754607667952,0.0,0.0,20.343347299270288,0.0,1.5353809597652572,1.6690857235268273e-14,0.0,1.4592465395800775e-15,
+2004ef,2453262.61,pW3,4104.6837,4514.733,10454.06632613563,0.0,0.0,98.8246855211191,0.0,3.605551678297641,7.118445257235789e-14,0.0,3.475711145646241e-15,
+2004ef,2453262.61,pW4,4508.833,5113.5821,19096.25189588555,0.0,0.0,162.3103438596992,0.0,2.989371507892129,8.804320449506235e-14,0.0,2.237482232858313e-15,
+2004ef,2453262.61,pW6,5656.3812,5889.4308,10517.722248570588,0.0,0.0,22.052335006473946,0.0,1.1081188506594022,8.10446397990869e-15,0.0,4.538275562986868e-16,
+2004ef,2453262.61,pW5,5184.3820000000005,5656.3812,8397.907254522093,0.0,0.0,86.33672392178543,0.0,2.472840342677646,3.604795134753736e-14,0.0,1.244149063580429e-15,
+2004ef,2453262.61,pW7,5877.6309,6364.3801,11661.337723724453,0.0,0.0,127.33378952878492,0.0,2.402975061200462,4.264558339805821e-14,0.0,1.0762735397296514e-15,
+2004ef,2453262.61,pW8,7812.8278,8500.1767,14143.14786110895,0.0,0.0,178.8799650460889,0.0,17.59997347924236,1.98416250313189e-14,0.0,2.618182103312865e-15,
+2004eo,2453268.51,pW2,3929.9058,4088.6087999999995,8676.400122496301,0.0,0.0,15.261708884344444,0.0,1.558833298378104,2.6710081943586488e-14,0.0,2.9821172428920573e-15,
